+++ EleCfitsioWrapper/EleCfitsioWrapper/impl/CompressionWrapper.hpp
@@ -9,22 +9,18 @@
 #include "EleCfitsioWrapper/HeaderWrapper.h"
 
 namespace Euclid {
+
 namespace Fits {
-namespace Compression {
 
 // used to dispatch the compress() call for each AlgoMixin subclass TDerived
 template <typename TDerived>
 void AlgoMixin<TDerived>::compress(void* fptr) const {
-  Euclid::Cfitsio::Compression::compress((fitsfile*)fptr, static_cast<const TDerived&>(*this));
+  Euclid::Cfitsio::compress((fitsfile*)fptr, static_cast<const TDerived&>(*this));
 }
 
-} // namespace Compression
 } // namespace Fits
-} // namespace Euclid
 
-namespace Euclid {
 namespace Cfitsio {
-namespace Compression {
 
 bool isCompressing(fitsfile* fptr) {
   int status = 0;
@@ -34,14 +30,14 @@ bool isCompressing(fitsfile* fptr) {
   return algo != int(NULL);
 }
 
-std::unique_ptr<Fits::Compression::Algo> readCompression(fitsfile* fptr) {
+std::unique_ptr<Fits::Algo> readCompression(fitsfile* fptr) {
 
   // Read algo
   int status = 0;
   int algo = int(NULL);
   fits_get_compression_type(fptr, &algo, &status);
   CfitsioError::mayThrow(status, fptr, "Cannot read compression type");
-  std::unique_ptr<Fits::Compression::Algo> out(new Fits::Compression::None());
+  std::unique_ptr<Fits::Algo> out(new Fits::None());
   if (algo == int(NULL)) {
     return out;
   }
@@ -58,23 +54,23 @@ std::unique_ptr<Fits::Compression::Algo> readCompression(fitsfile* fptr) {
   float f = 0;
   auto parametrize = [](float p) {
     if (p == 0) {
-      return Fits::Compression::Param::none();
+      return Fits::Param::none();
     } else if (p < 0) {
-      return Fits::Compression::Param::absolute(-p);
+      return Fits::Param::absolute(-p);
     } else {
-      return Fits::Compression::Param::relative(p);
+      return Fits::Param::relative(p);
     }
   };
   fits_get_quantize_level(fptr, &f, &status);
-  Fits::Compression::Quantization quantization(parametrize(f));
+  Fits::Quantization quantization(parametrize(f));
   if (quantization && HeaderIo::hasKeyword(fptr, "FZQMETHD")) {
     const std::string method = HeaderIo::parseRecord<std::string>(fptr, "FZQMETHD");
     if (method == "NO_DITHER") {
-      quantization.dithering(Fits::Compression::Dithering::None);
+      quantization.dithering(Fits::Dithering::None);
     } else if (method == "SUBTRACTIVE_DITHER_1") {
-      quantization.dithering(Fits::Compression::Dithering::EveryPixel);
+      quantization.dithering(Fits::Dithering::EveryPixel);
     } else if (method == "SUBTRACTIVE_DITHER_2") {
-      quantization.dithering(Fits::Compression::Dithering::NonZeroPixel);
+      quantization.dithering(Fits::Dithering::NonZeroPixel);
     } else {
       Fits::FitsError(std::string("Unknown compression dithering method: ") + method);
     }
@@ -89,24 +85,24 @@ std::unique_ptr<Fits::Compression::Algo> readCompression(fitsfile* fptr) {
 
   switch (algo) {
     case RICE_1:
-      out.reset(new Fits::Compression::Rice(tiling));
+      out.reset(new Fits::Rice(tiling));
       break;
     case HCOMPRESS_1:
-      out.reset(new Fits::Compression::HCompress({tiling[0], tiling[1]}));
+      out.reset(new Fits::HCompress({tiling[0], tiling[1]}));
       fits_get_hcomp_scale(fptr, &f, &status);
-      dynamic_cast<Fits::Compression::HCompress&>(*out).scale(std::move(scaling)).quantization(std::move(quantization));
+      dynamic_cast<Fits::HCompress&>(*out).scale(std::move(scaling)).quantization(std::move(quantization));
       break;
     case PLIO_1:
-      out.reset(new Fits::Compression::Plio(tiling));
-      dynamic_cast<Fits::Compression::Plio&>(*out).quantization(std::move(quantization));
+      out.reset(new Fits::Plio(tiling));
+      dynamic_cast<Fits::Plio&>(*out).quantization(std::move(quantization));
       break;
     case GZIP_1:
-      out.reset(new Fits::Compression::Gzip(tiling));
-      dynamic_cast<Fits::Compression::Gzip&>(*out).quantization(std::move(quantization));
+      out.reset(new Fits::Gzip(tiling));
+      dynamic_cast<Fits::Gzip&>(*out).quantization(std::move(quantization));
       break;
     case GZIP_2:
-      out.reset(new Fits::Compression::ShuffledGzip(tiling));
-      dynamic_cast<Fits::Compression::ShuffledGzip&>(*out).quantization(std::move(quantization));
+      out.reset(new Fits::ShuffledGzip(tiling));
+      dynamic_cast<Fits::ShuffledGzip&>(*out).quantization(std::move(quantization));
       break;
     default:
       throw Fits::FitsError("Unknown compression type");
@@ -123,12 +119,12 @@ inline void setTiling(fitsfile* fptr, const Fits::Position<-1>& shape) {
   CfitsioError::mayThrow(status, fptr, "Cannot set compression tiling");
 }
 
-inline void setQuantize(fitsfile* fptr, const Fits::Compression::Quantization& quantization) {
+inline void setQuantize(fitsfile* fptr, const Fits::Quantization& quantization) {
 
   int status = 0;
 
   // Set quantization level
-  if (quantization.level().type() == Fits::Compression::Param::Type::Absolute) {
+  if (quantization.level().type() == Fits::Param::Type::Absolute) {
     fits_set_quantize_level(fptr, -quantization.level().value(), &status);
     CfitsioError::mayThrow(status, fptr, "Cannot set absolute quantization level");
   } else { // None or relative quantization level
@@ -143,26 +139,26 @@ inline void setQuantize(fitsfile* fptr, const Fits::Compression::Quantization& q
   // Set dithering method
   // fits_set_quantize_method() is exact same as set_quantize_dither() (.._method() is the old name)
   switch (quantization.dithering()) {
-    case Fits::Compression::Dithering::EveryPixel:
+    case Fits::Dithering::EveryPixel:
       fits_set_quantize_dither(fptr, SUBTRACTIVE_DITHER_1, &status);
       break;
-    case Fits::Compression::Dithering::NonZeroPixel:
+    case Fits::Dithering::NonZeroPixel:
       fits_set_quantize_dither(fptr, SUBTRACTIVE_DITHER_2, &status);
       break;
-    case Fits::Compression::Dithering::None:
+    case Fits::Dithering::None:
       fits_set_quantize_dither(fptr, NO_DITHER, &status);
       break;
   }
   CfitsioError::mayThrow(status, fptr, "Cannot set dithering method");
 }
 
-void compress(fitsfile* fptr, const Fits::Compression::None&) {
+void compress(fitsfile* fptr, const Fits::None&) {
   int status = 0;
   fits_set_compression_type(fptr, int(NULL), &status);
   CfitsioError::mayThrow(status, fptr, "Cannot set compression type to None");
 }
 
-void compress(fitsfile* fptr, const Fits::Compression::Rice& algo) {
+void compress(fitsfile* fptr, const Fits::Rice& algo) {
   int status = 0;
   fits_set_compression_type(fptr, RICE_1, &status);
   CfitsioError::mayThrow(status, fptr, "Cannot set compression type to Rice");
@@ -170,7 +166,7 @@ void compress(fitsfile* fptr, const Fits::Compression::Rice& algo) {
   setQuantize(fptr, algo.quantization());
 }
 
-void compress(fitsfile* fptr, const Fits::Compression::HCompress& algo) {
+void compress(fitsfile* fptr, const Fits::HCompress& algo) {
 
   int status = 0;
 
@@ -180,7 +176,7 @@ void compress(fitsfile* fptr, const Fits::Compression::HCompress& algo) {
   setTiling(fptr, algo.tiling());
   setQuantize(fptr, algo.quantization());
 
-  if (algo.scale().type() == Fits::Compression::Param::Type::Absolute) {
+  if (algo.scale().type() == Fits::Param::Type::Absolute) {
     fits_set_hcomp_scale(fptr, -algo.scale().value(), &status);
     CfitsioError::mayThrow(status, fptr, "Cannot set absolute scale for HCompress");
   } else { // None or relative scaling applied in this case
@@ -192,14 +188,14 @@ void compress(fitsfile* fptr, const Fits::Compression::HCompress& algo) {
   CfitsioError::mayThrow(status, fptr, "Cannot set smoothing for HCompress");
 }
 
-void compress(fitsfile* fptr, const Fits::Compression::Plio& algo) {
+void compress(fitsfile* fptr, const Fits::Plio& algo) {
   int status = 0;
   fits_set_compression_type(fptr, PLIO_1, &status);
   CfitsioError::mayThrow(status, fptr, "Cannot set compression type to Plio");
   setTiling(fptr, algo.tiling());
 }
 
-void compress(fitsfile* fptr, const Fits::Compression::Gzip& algo) {
+void compress(fitsfile* fptr, const Fits::Gzip& algo) {
   int status = 0;
   fits_set_compression_type(fptr, GZIP_1, &status);
   CfitsioError::mayThrow(status, fptr, "Cannot set compression type to Gzip");
@@ -207,7 +203,7 @@ void compress(fitsfile* fptr, const Fits::Compression::Gzip& algo) {
   setQuantize(fptr, algo.quantization());
 }
 
-void compress(fitsfile* fptr, const Fits::Compression::ShuffledGzip& algo) {
+void compress(fitsfile* fptr, const Fits::ShuffledGzip& algo) {
   int status = 0;
   fits_set_compression_type(fptr, GZIP_2, &status);
   CfitsioError::mayThrow(status, fptr, "Cannot set compression type to ShuffledGzip");
@@ -215,7 +211,6 @@ void compress(fitsfile* fptr, const Fits::Compression::ShuffledGzip& algo) {
   setQuantize(fptr, algo.quantization());
 }
 
-} // namespace Compression
 } // namespace Cfitsio
 } // namespace Euclid