+++ EleFitsData/tests/src/Compression_test.cpp
@@ -10,14 +10,12 @@ using namespace Euclid::Fits;
 
 namespace Euclid {
 namespace Fits {
-namespace Compression {
 
 // Dummy definition to please compiler, although the method is not used here
 template <typename TDerived>
 void AlgoMixin<TDerived>::compress(void*) const {}
 // FIXME rm when the function is defined in Compression.hpp
 
-} // namespace Compression
 } // namespace Fits
 } // namespace Euclid
 
@@ -33,69 +31,69 @@ BOOST_AUTO_TEST_CASE(factor_test) {
   const float zeroFactor = 0.f;
   const float negativeFactor = -5.f;
 
-  const Compression::Param none_f = Compression::Param::none();
+  const Param none_f = Param::none();
   BOOST_TEST(none_f.value() == zeroFactor);
-  BOOST_TEST(none_f.type() == Compression::Param::Type::None);
+  BOOST_TEST(none_f.type() == Param::Type::None);
 
-  const Compression::Param absolute_f = Compression::Param::absolute(positiveFactor);
-  BOOST_CHECK_THROW(Compression::Param::absolute(zeroFactor), Euclid::Fits::FitsError);
-  BOOST_CHECK_THROW(Compression::Param::absolute(negativeFactor), Euclid::Fits::FitsError);
+  const Param absolute_f = Param::absolute(positiveFactor);
+  BOOST_CHECK_THROW(Param::absolute(zeroFactor), Euclid::Fits::FitsError);
+  BOOST_CHECK_THROW(Param::absolute(negativeFactor), Euclid::Fits::FitsError);
 
   BOOST_TEST(absolute_f.value() == positiveFactor);
-  BOOST_TEST(absolute_f.type() == Compression::Param::Type::Absolute);
+  BOOST_TEST(absolute_f.type() == Param::Type::Absolute);
 
-  const Compression::Param relative_f = Compression::Param::relative(positiveFactor);
-  BOOST_CHECK_THROW(Compression::Param::relative(zeroFactor), Euclid::Fits::FitsError);
-  BOOST_CHECK_THROW(Compression::Param::relative(negativeFactor), Euclid::Fits::FitsError);
+  const Param relative_f = Param::relative(positiveFactor);
+  BOOST_CHECK_THROW(Param::relative(zeroFactor), Euclid::Fits::FitsError);
+  BOOST_CHECK_THROW(Param::relative(negativeFactor), Euclid::Fits::FitsError);
 
   BOOST_TEST(relative_f.value() == positiveFactor);
-  BOOST_TEST(relative_f.type() == Compression::Param::Type::Relative);
+  BOOST_TEST(relative_f.type() == Param::Type::Relative);
 
   // testing == operator:
-  const Compression::Param f1 = Compression::Param::relative(5.f);
-  const Compression::Param f2 = Compression::Param::relative(5.f);
-  const Compression::Param f3 = Compression::Param::absolute(5.f);
-  const Compression::Param f4 = Compression::Param::relative(4.f);
+  const Param f1 = Param::relative(5.f);
+  const Param f2 = Param::relative(5.f);
+  const Param f3 = Param::absolute(5.f);
+  const Param f4 = Param::relative(4.f);
   BOOST_TEST((f1 == f2));
   BOOST_TEST(not(f1 == f3));
   BOOST_TEST(not(f1 == f4));
 }
 
 BOOST_AUTO_TEST_CASE(default_quantization_test) {
-  Compression::Quantization q;
+  Quantization q;
   BOOST_TEST(not q);
   BOOST_TEST(not q.level());
-  BOOST_TEST((q.dithering() == Compression::Dithering::None));
-  BOOST_CHECK_THROW(q.dithering(Compression::Dithering::EveryPixel), FitsError); // Cannot dither disabled q
+  BOOST_TEST((q.dithering() == Dithering::None));
+  BOOST_CHECK_THROW(q.dithering(Dithering::EveryPixel), FitsError); // Cannot dither disabled q
 }
 
 BOOST_AUTO_TEST_CASE(default_dithering_test) {
-  const auto level = Compression::Param::absolute(4); // CFITSIO default
-  Compression::Quantization q(level);
+  const auto level = Param::absolute(4); // CFITSIO default
+  Quantization q(level);
   BOOST_TEST(q);
   BOOST_TEST(q.level() == level);
-  BOOST_TEST((q.dithering() == Compression::Dithering::EveryPixel));
+  BOOST_TEST((q.dithering() == Dithering::EveryPixel));
 }
 
 BOOST_AUTO_TEST_CASE(quantization_level_test) {
-  const auto level = Compression::Param::relative(4);
-  Compression::Quantization q;
+  const auto level = Param::relative(4);
+  Quantization q;
   q.level(level);
   BOOST_TEST((q.level() == level));
-  BOOST_TEST((q.dithering() == Compression::Dithering::None));
-  q.dithering(Compression::Dithering::NonZeroPixel);
+  BOOST_TEST((q.dithering() == Dithering::None));
+  q.dithering(Dithering::NonZeroPixel);
   BOOST_TEST((q.level() == level));
-  BOOST_TEST((q.dithering() == Compression::Dithering::NonZeroPixel));
+  BOOST_TEST((q.dithering() == Dithering::NonZeroPixel));
 }
 
 BOOST_AUTO_TEST_CASE(quantization_equality_test) {
-  Compression::Quantization q0;
-  Compression::Quantization q0n(Compression::Param::none(), Compression::Dithering::None);
-  Compression::Quantization q3(Compression::Param::absolute(3));
-  Compression::Quantization q4(Compression::Param::absolute(4));
-  Compression::Quantization q4n(Compression::Param::absolute(4), Compression::Dithering::None);
-  Compression::Quantization q4e(Compression::Param::absolute(4), Compression::Dithering::EveryPixel);
-  Compression::Quantization q4nz(Compression::Param::absolute(4), Compression::Dithering::NonZeroPixel);
+  Quantization q0;
+  Quantization q0n(Param::none(), Dithering::None);
+  Quantization q3(Param::absolute(3));
+  Quantization q4(Param::absolute(4));
+  Quantization q4n(Param::absolute(4), Dithering::None);
+  Quantization q4e(Param::absolute(4), Dithering::EveryPixel);
+  Quantization q4nz(Param::absolute(4), Dithering::NonZeroPixel);
   BOOST_TEST((q0 == q0n));
   BOOST_TEST((q0 != q4n));
   BOOST_TEST((q3 != q4));
@@ -116,21 +114,21 @@ void testAlgoMixinParameters(long dimension = 0) {
   BOOST_TEST((shape2 == shape));
 
   // check default quantization values:
-  BOOST_TEST((algo.quantization() == Compression::Quantization()));
+  BOOST_TEST((algo.quantization() == Quantization()));
 
   // set/get quantization:
-  Compression::Quantization quantization;
-  quantization.level(Compression::Param::absolute(5));
-  quantization.dithering(Compression::Dithering::None);
+  Quantization quantization;
+  quantization.level(Param::absolute(5));
+  quantization.dithering(Dithering::None);
   algo.quantization(quantization);
   BOOST_TEST((algo.quantization() == quantization));
 }
 
 // specific to the None algo
 template <>
-void testAlgoMixinParameters<Compression::None>(long) {
+void testAlgoMixinParameters<None>(long) {
 
-  Compression::None algo;
+  None algo;
 
   // verify shape of algo is correctly stored at construction
   BOOST_TEST((algo.tiling() == Position<-1>()));
@@ -138,54 +136,54 @@ void testAlgoMixinParameters<Compression::None>(long) {
 
 BOOST_AUTO_TEST_CASE(algo_mixin_test) {
 
-  testAlgoMixinParameters<Compression::None>();
+  testAlgoMixinParameters<None>();
 
-  testAlgoMixinParameters<Compression::Rice>(0);
-  testAlgoMixinParameters<Compression::Rice>(1);
-  testAlgoMixinParameters<Compression::Rice>(2);
-  testAlgoMixinParameters<Compression::Rice>(3);
-  testAlgoMixinParameters<Compression::Rice>(4);
-  testAlgoMixinParameters<Compression::Rice>(5);
-  testAlgoMixinParameters<Compression::Rice>(6);
+  testAlgoMixinParameters<Rice>(0);
+  testAlgoMixinParameters<Rice>(1);
+  testAlgoMixinParameters<Rice>(2);
+  testAlgoMixinParameters<Rice>(3);
+  testAlgoMixinParameters<Rice>(4);
+  testAlgoMixinParameters<Rice>(5);
+  testAlgoMixinParameters<Rice>(6);
 
-  testAlgoMixinParameters<Compression::HCompress>(2);
+  testAlgoMixinParameters<HCompress>(2);
 
-  testAlgoMixinParameters<Compression::Plio>(0);
-  testAlgoMixinParameters<Compression::Plio>(1);
-  testAlgoMixinParameters<Compression::Plio>(2);
-  testAlgoMixinParameters<Compression::Plio>(3);
-  testAlgoMixinParameters<Compression::Plio>(4);
-  testAlgoMixinParameters<Compression::Plio>(5);
-  testAlgoMixinParameters<Compression::Plio>(6);
+  testAlgoMixinParameters<Plio>(0);
+  testAlgoMixinParameters<Plio>(1);
+  testAlgoMixinParameters<Plio>(2);
+  testAlgoMixinParameters<Plio>(3);
+  testAlgoMixinParameters<Plio>(4);
+  testAlgoMixinParameters<Plio>(5);
+  testAlgoMixinParameters<Plio>(6);
 
-  testAlgoMixinParameters<Compression::Gzip>(0);
-  testAlgoMixinParameters<Compression::Gzip>(1);
-  testAlgoMixinParameters<Compression::Gzip>(2);
-  testAlgoMixinParameters<Compression::Gzip>(3);
-  testAlgoMixinParameters<Compression::Gzip>(4);
-  testAlgoMixinParameters<Compression::Gzip>(5);
-  testAlgoMixinParameters<Compression::Gzip>(6);
+  testAlgoMixinParameters<Gzip>(0);
+  testAlgoMixinParameters<Gzip>(1);
+  testAlgoMixinParameters<Gzip>(2);
+  testAlgoMixinParameters<Gzip>(3);
+  testAlgoMixinParameters<Gzip>(4);
+  testAlgoMixinParameters<Gzip>(5);
+  testAlgoMixinParameters<Gzip>(6);
 
-  testAlgoMixinParameters<Compression::ShuffledGzip>(0);
-  testAlgoMixinParameters<Compression::ShuffledGzip>(1);
-  testAlgoMixinParameters<Compression::ShuffledGzip>(2);
-  testAlgoMixinParameters<Compression::ShuffledGzip>(3);
-  testAlgoMixinParameters<Compression::ShuffledGzip>(4);
-  testAlgoMixinParameters<Compression::ShuffledGzip>(5);
-  testAlgoMixinParameters<Compression::ShuffledGzip>(6);
+  testAlgoMixinParameters<ShuffledGzip>(0);
+  testAlgoMixinParameters<ShuffledGzip>(1);
+  testAlgoMixinParameters<ShuffledGzip>(2);
+  testAlgoMixinParameters<ShuffledGzip>(3);
+  testAlgoMixinParameters<ShuffledGzip>(4);
+  testAlgoMixinParameters<ShuffledGzip>(5);
+  testAlgoMixinParameters<ShuffledGzip>(6);
 }
 
 BOOST_AUTO_TEST_CASE(hcompress_test) {
 
   const Position<-1>& shape {300, 200};
-  Compression::HCompress algo(shape);
+  HCompress algo(shape);
 
   // check default hcompress params values:
-  BOOST_TEST((algo.scale() == Compression::Param::none()));
+  BOOST_TEST((algo.scale() == Param::none()));
   BOOST_TEST(algo.isSmooth() == false);
 
   // setters & getters:
-  Compression::Param scale = Compression::Param::absolute(5.f);
+  Param scale = Param::absolute(5.f);
   algo.scale(scale);
   algo.enableSmoothing();
   BOOST_TEST((algo.scale() == scale));